agent/requirements.txt
antigravity
groq
google-genai
openai